# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Todo Application RES-API scheme",
        "title": "Golang Todo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5050",
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks": {
            "get": {
                "description": "Список задач с пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Список задач",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Фильтрация по ID автора задачи",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Размер страницы с задачами",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение страницы с задачами",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_features_tasks_transport_http.TaskDTOResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создать новую задачу в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Создать задачу",
                "parameters": [
                    {
                        "description": "CreateTask тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешно созданная задача",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Author not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "description": "Получение задачи",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получение задачи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача успешно найдена",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.GetTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удалить задачу в системе по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Удалить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID удаляемой задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешное удаление задачи"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Изменение информации существующей задачи\n### Логика обновления полей (Three-state logic)\n1. **Поле не передано**: `description` игнорируется, значение в БД не меняется\n2. **Явно передано значение**: `\"description\": \"Тест задания\"` - устанавливает новый номер телефона\n3. **Передан null**: `\"description\": null` - очищается поле в БО (set to NULL)\nОграничения: `title` и `completed` не могут выставлены как null",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Изменить задачу",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID задачи",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PatchTask тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.PatchTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно изменненая задача",
                        "schema": {
                            "$ref": "#/definitions/internal_features_tasks_transport_http.PatchTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "Список пользователей с пагинацией",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Список пользователей",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Размер страницы с пользователями",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение страницы с пользователями",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список пользователей",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_features_users_transport_http.UserDTOResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Создать нового пользователя в системе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Создать пользователя",
                "parameters": [
                    {
                        "description": "CreateUser тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Успешно созданный пользователь",
                        "schema": {
                            "$ref": "#/definitions/internal_features_users_transport_http.CreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_mihail2771_todogo_internal_core_transport_http_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "full error text"
                },
                "message": {
                    "type": "string",
                    "example": "short readable message"
                }
            }
        },
...
        "internal_features_tasks_transport_http.CreateTaskRequest": {
            "type": "object",
            "required": [
                "author_user_id",
                "title"
            ],
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 2
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1,
                    "example": "Описание задачи"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1,
                    "example": "Название задачи"
                }
            }
        },
        "internal_features_tasks_transport_http.CreateTaskResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 2
                },
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "completed_at": {
                    "type": "string",
                    "example": "null"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Описание задачи"
                },
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "title": {
                    "type": "string",
                    "example": "Название задачи"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "internal_features_tasks_transport_http.GetTaskResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 2
                },
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "completed_at": {
                    "type": "string",
                    "example": "null"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Описание задачи"
                },
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "title": {
                    "type": "string",
                    "example": "Название задачи"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "internal_features_tasks_transport_http.PatchTaskRequest": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 2
                },
                "completed": {
                    "type": "boolean",
                    "example": true
                },
                "description": {
                    "type": "string",
                    "example": "Описание задачи"
                },
                "title": {
                    "type": "string",
                    "example": "Название задачи"
                }
            }
        },
        "internal_features_tasks_transport_http.PatchTaskResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 2
                },
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "completed_at": {
                    "type": "string",
                    "example": "null"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Описание задачи"
                },
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "title": {
                    "type": "string",
                    "example": "Название задачи"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "internal_features_tasks_transport_http.TaskDTOResponse": {
            "type": "object",
            "properties": {
                "author_user_id": {
                    "type": "integer",
                    "example": 2
                },
                "completed": {
                    "type": "boolean",
                    "example": false
                },
                "completed_at": {
                    "type": "string",
                    "example": "null"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-02-26T10:30:00Z"
                },
                "description": {
                    "type": "string",
                    "example": "Описание задачи"
                },
                "id": {
                    "type": "integer",
                    "example": 15
                },
                "title": {
                    "type": "string",
                    "example": "Название задачи"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
        "internal_features_users_transport_http.CreateUserRequest": {
            "type": "object",
            "required": [
                "full_name"
            ],
            "properties": {
                "full_name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Ivan Ivanov"
                },
                "phone_number": {
                    "type": "string",
                    "maxLength": 15,
                    "minLength": 10,
                    "example": "+79994412345"
                }
            }
        },
        "internal_features_users_transport_http.CreateUserResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Ivan Ivanov"
                },
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79994412345"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
        "internal_features_users_transport_http.UserDTOResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string",
                    "example": "Ivan Ivanov"
                },
                "id": {
                    "type": "integer",
                    "example": 10
                },
                "phone_number": {
                    "type": "string",
                    "example": "+79994412345"
                },
                "version": {
                    "type": "integer",
                    "example": 3
                }
            }
        }
    }
}